requirements.txt
flask==3.0.2
werkzeug==3.0.1
PyPDF2==3.0.1
google-generativeai==0.3.2
python-dotenv==1.0.1
gunicorn==21.2.0
google-cloud-documentai==2.25.0
google-cloud-storage==2.15.0
google-cloud-logging==3.9.0 